[AMORO-3559] Extract abstract class for table metrics
